NF: remove useless casts

+++ AnkiDroid/src/main/java/com/ichi2/libanki/sched/SchedV2.java
@@ -1229,7 +1229,7 @@ public class SchedV2 extends AbstractSched {
         // due today?
         if (card.getDue() < mDayCutoff) {
             // Add some randomness, up to 5 minutes or 25%
-            int maxExtra = (int) Math.min(300, (int)(delay * 0.25));
+            int maxExtra = Math.min(300, (int)(delay * 0.25));
             int fuzz = new Random().nextInt(maxExtra);
             card.setDue(Math.min(mDayCutoff - 1, card.getDue() + fuzz));
             card.setQueue(Consts.QUEUE_TYPE_LRN);
